GUI now functioning again and minor Event refactor
The GUI is now almost back to what it was now in JavaFX, still has a
weird width bug, but does fully function. I also did some refactoring
with the Event class and its subclasses to get rid of some redundant
code and make it easier to access an event's list of schools.